+++ src/components/create_hackathon/CreateHackathonPage.js
@@ -7,9 +7,9 @@ import CheckCircleIcon from "@material-ui/icons/CheckCircle";
 import SaveIcon from "@material-ui/icons/Save";
 import DeleteIcon from "@material-ui/icons/Delete";
 import OverviewEditor from "../hack_forms/overview/OverviewEditor";
-import HackathonDetailsForm from "../hack_forms/HackathonDetailsForm";
+import HackathonDetailsForm from "../hack_forms/details/HackathonDetailsForm";
 import { DASHBOARD_ROUTE } from "../../routes";
-import RegistrationDetailsForm from "../hack_forms/RegistrationDetailsForm";
+import RegistrationDetailsForm from "../hack_forms/questions/RegistrationDetailsForm";
 import { QUESTION_TYPE } from "../hack_forms/questions/QuestionType";
 import { PAGES, PAGE_TITLES, PAGES_LIST } from "./CreateHackathonSubpages";
 import HackathonPreviewForm from "../hack_forms/HackathonPreviewForm";
